+++ scirs2-core/src/resource/storage.rs
@@ -25,8 +25,8 @@ impl Default for StorageInfo {
             devices: vec![StorageDevice::default()],
             optimal_io_size: 64 * 1024, // 64KB
             queue_depth: 32,
-            capacity: 500 * 1024 * 1024 * 1024,  // 500GB
-            available: 250 * 1024 * 1024 * 1024, // 250GB
+            capacity: (500u64 * 1024 * 1024 * 1024) as usize, // 500GB
+            available: (250u64 * 1024 * 1024 * 1024) as usize, // 250GB
         }
     }
 }
@@ -73,8 +73,8 @@ impl Default for StorageDevice {
             device_type: StorageType::Ssd,
             is_ssd: true,
             supports_async: true,
-            capacity: 500 * 1024 * 1024 * 1024, // 500GB
-            optimal_io_size: 64 * 1024,         // 64KB
+            capacity: (500u64 * 1024 * 1024 * 1024) as usize, // 500GB
+            optimal_io_size: 64 * 1024,                       // 64KB
         }
     }
 }
